requirements.txt
streamlit>=1.28.0
pdfplumber>=0.9.0
PyMuPDF>=1.23.0
spacy>=3.6.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
pandas>=2.0.0
numpy>=1.24.0
fpdf2>=2.7.0
rapidfuzz>=3.5.0
plotly>=5.15.0
